Update gmic-community repository URL.
It appears that the gmic-community repository is now owned by GreycLab
instead of dtschump. This change updates the URL used to fetch the
source so we don't have to rely on GitHub redirecting anymore. It also
deletes a bunch of commented out rules that don't appear to be needed
anymore.